IC-1: fix: fix struct of types in files
